+++ pkg/env/sensor.go
@@ -37,4 +37,9 @@ var (
 
 	// ScanTimeout defines the scan timeout duration for Sensor initiated scans
 	ScanTimeout = registerDurationSetting("ROX_SCAN_TIMEOUT", 6*time.Minute)
+
+	// DelegatedScanningDisabled disables the capabilities associated with delegated image scanning.
+	// This is meant to be a 'kill switch' that allows for local scanning to continue (ie: for OCP internal repos)
+	// in the event the delegated scanning capabilities are causing unforeseen issues.
+	DelegatedScanningDisabled = RegisterBooleanSetting("ROX_DELEGATED_SCANNING_DISABLED", false)
 )